added path to slurm dependencies lib
If libslurm.so.40 is not in a dedicated path such as /var/lib/slurm-llnl but in a generic one like /usr/lib/x86_64-linux-gnu/ binding this latter one will crash with container libraries. But if we use just /usr/lib/x86_64-linux-gnu/libslurm.so.*, there will be some missing dependencies. Here we bind them too.